Refactor galaxy.util.handlers -> galaxy.web.stack.handlers.
Nothing in galaxy.util should import sqlalchemy stuff and this package does after #7129.
