+++ lib/galaxy/workflow/scheduling_manager.py
@@ -7,8 +7,8 @@ import galaxy.workflow.schedulers
 from galaxy import model
 from galaxy.exceptions import HandlerAssignmentError
 from galaxy.util import plugin_config
-from galaxy.util.handlers import ConfiguresHandlers, HANDLER_ASSIGNMENT_METHODS
 from galaxy.util.monitors import Monitors
+from galaxy.web.stack.handlers import ConfiguresHandlers, HANDLER_ASSIGNMENT_METHODS
 from galaxy.web.stack.message import WorkflowSchedulingMessage
 
 log = logging.getLogger(__name__)
